+++ CHANGELOG.md
@@ -22,6 +22,7 @@ however, insignificant breaking changes do not guarantee a major version bump, s
 - Added custom database name detection in `CONNECTION_URI`
 - Added dpy and python version info in `?about` embed and credit for contributors
 - New config option: `mention_message`, this will include an additional configurable message next to role mention on thread creation. Defaults to None.
+- Support for trailing space in `?prefix` command, `?prefix "mm "
 
 ### Changed
 - Guild icons in embed footers and author urls now have a fixed size of 128. ([PR #3261](https://github.com/modmail-dev/modmail/pull/3261))